Move SectionHeaderConfigurator into a separate class
So that other kinds of configurators can be passed in to
SectionedListAdapter.
